Update: align test files with the source changes in this PR
