+++ test/structure/structure_region/test_capping.py
@@ -22,7 +22,7 @@ DATA_DIR = f"{CURR_DIR}/../data/"
 WORK_DIR = f"{CURR_DIR}/../work_dir/"
 sp = PDBParser()
 
-def test_capping_with_residue_terminals(helpers):
+def test_capping_with_residue_terminals_tri_a(helpers):
     """as name. use an example region from selection.
     Just make sure not execptions is raised for now.
     TODO complete this test"""
@@ -226,19 +226,3 @@ def test_capping_with_residue_terminals(helpers):
 
     fs.safe_rm(test_file)
 
-def test_residue_cap_deepcopy():
-    """test the behavior of deepcopy for residue cap"""
-    test_stru = sp.get_structure(f"{DATA_DIR}KE_07_R7_2_S.pdb")
-    test_res = test_stru.residues[0]
-    test_res_2 = test_stru.residues[1]
-    test_ch3 = CH3Cap(
-        test_res,
-        test_res.find_atom_name('C'),
-        test_res_2.find_atom_name('N'),
-        'cterm' 
-    )
-    ch3_copy = deepcopy(test_ch3)
-
-    assert ch3_copy.link_residue is not test_res
-    assert ch3_copy.link_residue.parent is None
-    assert ch3_copy.link_atom in ch3_copy.link_residue.atoms
